Split out authors
